backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
torch==2.0.1+cu118
torchvision==0.15.2+cu118
Pillow==9.5.0
numpy==1.24.3
